requirements.txt
absl-py==2.1.0
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==23.11.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cycler==0.12.1
deepface==0.0.89
dill==0.3.7
Django==4.2.7
django-admin-multiple-choice-list-filter==0.1.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-filter==24.2
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.3
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
inflection==0.5.1
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
pathspec==0.11.2
phonenumberslite==8.13.25
pillow==10.2.0
platformdirs==4.0.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
retina-face==0.0.15
rich==13.7.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tomli==2.0.1
tomlkit==0.12.3
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
